Another round of comments addressed
